pyproject.toml
[project]
name = "imagetotext-python-azure-ai-opensource-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "azure-cognitiveservices-vision-computervision>=0.9.1",
    "python-dotenv>=1.2.1",
]
